Migrate VoiceMessageInput comp to have only UI functions
